Refactor bot.py into directory
